Refactor profile-user-core code to use profile-user-commons entities and utilities
