+++ airavata-services/profile-service/profile-user-core/src/main/java/org/apache/airavata/service/profile/user/core/exceptions/UserProfileRegistryException.java
@@ -18,22 +18,11 @@
  * under the License.
  *
 */
-package org.apache.airavata.service.profile.user.core.utils;
+package org.apache.airavata.service.profile.user.core.exceptions;
 
-import org.dozer.DozerBeanMapper;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-public class ObjectMapperSingleton extends DozerBeanMapper{
-    private final static Logger logger = LoggerFactory.getLogger(ObjectMapperSingleton.class);
-
-    private static ObjectMapperSingleton instance;
-
-    private ObjectMapperSingleton(){}
-
-    public static ObjectMapperSingleton getInstance(){
-        if(instance == null)
-            instance = new ObjectMapperSingleton();
-        return instance;
-    }
+public class UserProfileRegistryException extends Exception {
+    private final static Logger logger = LoggerFactory.getLogger(UserProfileRegistryException.class);
 }
